[Refactor] #102 StopConfirmationView 분리
